app/requirements.txt
Flask==2.0.2
gunicorn==20.1.0
#google-api-python-client==2.52.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
requests